# docs/swagger.yaml
basePath: /
definitions:
...
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  domain.UserProfile:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        type: string
      subcribers:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      subscribers_count:
        type: integer
      subscriptions:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      subscriptions_count:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      videos:
        items:
          $ref: '#/definitions/domain.Video'
        type: array
    type: object
  domain.Video:
    properties:
      author_id:
        type: integer
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      qualities:
        items:
          $ref: '#/definitions/domain.VideoQuality'
        type: array
      status:
        type: string
      thumbnail_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_url:
        type: string
      views:
        type: integer
    type: object
  domain.VideoQuality:
    properties:
      created_at:
        type: string
      id:
        type: integer
      quality:
        type: string
      video_id:
        type: integer
      video_url:
        type: string
    type: object
...
  httpdelivery.CommonError:
    properties:
      error:
        type: string
    type: object
  httpdelivery.CommonResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: analogyoutube@example.com
    name: AnalogYouTube Backend
  description: |-
    API для AnalogYouTube: пользователи, видео, чаты, донаты и остальное.
    Для закрытых запросов нажмите Authorize и вставьте: Bearer <token>.
  title: AnalogYouTube API
  version: "1.0"
paths:
  /api/admin/videos/{id}:
    delete:
      description: Полностью удаляет видео.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Админское удаление
      tags:
      - Admin
      - Videos
...
  /api/me:
    get:
      description: Данные текущего пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Мой профиль
      tags:
      - Profile
    put:
      consumes:
      - multipart/form-data
      description: Меняет мои данные, аватарку или пароль.
      parameters:
      - description: Новое имя пользователя
        in: formData
        name: username
        required: true
        type: string
      - description: Email, например user@gmail.com
        in: formData
        name: email
        type: string
      - description: Новый пароль
        in: formData
        name: password
        type: string
      - description: Описание канала
        in: formData
        name: description
        type: string
      - description: Файл аватарки
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить профиль
      tags:
      - Profile
...
  /api/users/{id}:
    get:
      description: Публичная страница канала.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserProfile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Профиль пользователя
      tags:
      - Profile
...
  /api/users/{id}/videos:
    get:
      description: Ролики выбранного автора.
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Video'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Видео пользователя
      tags:
      - Profile
  /api/videos:
    get:
      description: Видео для главной ленты.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Video'
            type: array
      summary: Рекомендации
      tags:
      - Videos
    post:
      consumes:
      - multipart/form-data
      description: Загружает видеофайл, обложку и данные ролика.
      parameters:
      - description: Название видео
        in: formData
        name: title
        required: true
        type: string
      - description: Описание видео
        in: formData
        name: description
        type: string
      - description: Имя категории
        in: formData
        name: category_name
        type: string
      - description: Файл видео
        in: formData
        name: video
        required: true
        type: file
      - description: Файл обложки
        in: formData
        name: thumbnail
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Video'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Создать видео
      tags:
      - Videos
  /api/videos/{id}:
    delete:
      description: Скрывает видео и оставляет архивную копию.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpdelivery.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Удалить видео
      tags:
      - Videos
    get:
      description: Данные видео и новый просмотр.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Video'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Видео по ID
      tags:
      - Videos
    put:
      consumes:
      - multipart/form-data
      description: Меняет данные ролика без замены видеофайла.
      parameters:
      - description: ID видео
        in: path
        name: id
        required: true
        type: integer
      - description: Название видео
        in: formData
        name: title
        required: true
        type: string
      - description: Описание видео
        in: formData
        name: description
        type: string
      - description: Имя категории
        in: formData
        name: category_name
        type: string
      - description: Новая обложка
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Video'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить видео
      tags:
      - Videos
...
  /api/videos/allvideos:
    get:
      description: Админский список всех видео.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Video'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      security:
      - BearerAuth: []
      summary: Все видео
      tags:
      - Admin
      - Videos
...
  /api/videos/search:
    get:
      description: Ищет ролики по названию.
      parameters:
      - description: Часть названия видео
        in: query
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Video'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpdelivery.CommonError'
      summary: Поиск видео
      tags:
      - Videos
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"